5.8 Role użytkowników

+++ Blogger/WebAPI/Controllers/V1/PostsController.cs
@@ -1,13 +1,11 @@
 ﻿using Application.Dto;
 using Application.Interfaces;
-using Azure;
-using Cosmonaut.Extensions;
-using Cosmonaut.Response;
-using Domain.Entities;
+using Infrastructure.Identity;
 using Microsoft.AspNet.OData;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Hosting;
 using Swashbuckle.AspNetCore.Annotations;
+using System.Security.Claims;
 using WebAPI.Filters;
 using WebAPI.Helpers;
 using WebAPI.Wrappers;
@@ -17,6 +15,7 @@ namespace WebAPI.Controllers.V1;
 
 [Route("api/[controller]")]
 [ApiVersion("1.0")]
+[Authorize()]
 [ApiController]
 public class PostsController : ControllerBase
 {
@@ -35,6 +34,7 @@ public class PostsController : ControllerBase
 
     [SwaggerOperation(Summary = "Retrieves all posts")]
     [EnableQuery]
+    [Authorize(Roles = UserRoles.Admin)]
     [HttpGet("[action]")]
     public async Task<IActionResult> Get([FromQuery] PaginationFilter paginationFilter, [FromQuery] SortingFilter sortingFilter, [FromQuery] string filterBy = "")
     {
@@ -49,6 +49,7 @@ public class PostsController : ControllerBase
     }
 
     [SwaggerOperation(Summary = "Retrieves a specific post by unique Id")]
+    [AllowAnonymous]
     [HttpGet("{id}")]
     public async Task<IActionResult> Get(int id)
     {
@@ -62,25 +63,40 @@ public class PostsController : ControllerBase
     }
 
     [SwaggerOperation(Summary = "Create a new post")]
+    [Authorize(Roles = UserRoles.User)]
     [HttpPost]
     public async Task<IActionResult> Create(CreatePostDto newPost)
     {
-        var post = await _postService.AddNewPostAsync(newPost);
+        var post = await _postService.AddNewPostAsync(newPost, User.FindFirstValue(ClaimTypes.NameIdentifier));
         return Created($"api/posts/{post.Id}", new Wrappers.Response<PostDto>(post));
     }
 
     [SwaggerOperation(Summary = "Update a exsisting post")]
+    [Authorize(Roles = UserRoles.User)]
     [HttpPut]
     public async Task<IActionResult> Update(UpdatePostDto updatePost)
     {
+        var userOwnsPost = await _postService.UserOwnsPostAsync(updatePost.Id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+        if (!userOwnsPost)
+        {
+            return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+        }
+
         await _postService.UpdatePostAsync(updatePost);
         return NoContent();
     }
 
     [SwaggerOperation(Summary = "Delete a specific post")]
+    [Authorize(Roles = UserRoles.AdminOrUser)]
     [HttpDelete("Id")]
     public async Task<IActionResult> Delete(int id)
     {
+        var userOwnsPost = await _postService.UserOwnsPostAsync(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+        var isAdmin = User.FindFirstValue(ClaimTypes.Role).Contains(UserRoles.Admin);
+        if (!isAdmin && !userOwnsPost)
+        {
+            return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+        }
         await _postService.DeletePostAsync(id);
         return NoContent();
     }
